long long is not the same as int64_t :o

+++ temp.cpp
@@ -9,7 +9,7 @@ using namespace std;
 #endif
 #define all(x) begin(x), end(x)
 
-using ll = int64_t;
+using ll = long long;
 using ld = long double;
 using pii = pair<int32_t, int32_t>;
 using vi = vector<int32_t>;
